1.Berhasil menjalankan GET PACTH DELETE faskes dimana patch dan delete memakai id yg didapat dari params id

+++ dtokemkes/app/Controllers/Http/FaskesController.ts
@@ -1,5 +1,6 @@
 import type { HttpContextContract } from '@ioc:Adonis/Core/HttpContext'
 import Faske from 'App/Models/Faske'
+import FaskesEditValidator from 'App/Validators/FaskesEditValidator'
 import FaskeValidator from 'App/Validators/FaskeValidator'
 
 export default class FaskesController {
@@ -21,7 +22,7 @@ export default class FaskesController {
         }
     }
     public async edit({ params, request, response }: HttpContextContract) {
-        const input = await request.validate(FaskeValidator)
+        const input = await request.validate(FaskesEditValidator)
         try {
             const faskes = await Faske.findBy('id', params.id)
             faskes?.merge(input)
@@ -34,8 +35,13 @@ export default class FaskesController {
     public async delete({ params, response }: HttpContextContract) {
         try {
             const faskes = await Faske.findBy('id', params.id)
-            await faskes?.delete()
-            return response.status(200).json({ code: 200, status: 'success', data: faskes })
+            console.log(faskes)
+            if(faskes){
+                await faskes?.delete()
+                return response.status(200).json({ code: 200, status: 'success', data: faskes })
+            }else{
+                return response.status(404).json({ code: 404, status: 'not found'})
+            }
         } catch (err) {
             return response.status(500).json({ code: 500, status: 'error', message: err.message })
         }
